Change boards/hc/mcs92s12ne64 to boards/hc/m9s12
fix the follow error:
Refresh demo9s12ne64/ostest
LN: include/arch/board to /workspace/mirtos/nuttx/boards/hc/mcs92s12ne64/demo9s12ne64/include
LN: include/arch/chip to arch/hc/include/mcs92s12ne64
No directory at /workspace/mirtos/nuttx/arch/hc/include/mcs92s12ne64
