+++ Documentation/NuttX.html
@@ -6400,7 +6400,7 @@ Mem:         29232       5920      23312      23312
          However, testing has not yet begun due to issues with BDMs, Code Warrior, and
          the paging in the build process.
          Progress is slow, but I hope to see a fully verified MC9S12NE64 port in the near future.
-         Refer to the NuttX board README files for <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/mcs92s12ne64/demo9s12ne64/README.txt" target="_blank">DEMO9S12NE64</a> and for the <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/mcs92s12ne64/ne64badge/README.txt" target="_blank">NE64 /PoE Badge</a> for further information.
+         Refer to the NuttX board README files for <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/m9s12/demo9s12ne64/README.txt" target="_blank">DEMO9S12NE64</a> and for the <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/m9s12/ne64badge/README.txt" target="_blank">NE64 /PoE Badge</a> for further information.
       </p>
     </ul>
   </td>
