+++ Documentation/README.html
@@ -409,11 +409,11 @@ nuttx/
  |   |       `- moteino-mega/
  |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/avr/atmega/moteino-mega/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |- hc/
- |   |   `- mcs92s12ne64/
+ |   |   `- m9s12/
  |   |       |- demo9s12ne64/
- |   |       |`  `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/mcs92s12ne64/demo9s12ne64/README.txt" target="_blank"><b><i>README.txt</i></b></a>
+ |   |       |`  `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/m9s12/demo9s12ne64/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |       `- ne64badge/
- |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/mcs92s12ne64/ne64badge/README.txt" target="_blank"><b><i>README.txt</i></b></a>
+ |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/hc/m9s12/ne64badge/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |- mips/
  |   |   |- pic32mx/
  |   |   |   |- mirtoo/
