+++ boards/hc/m9s12/ne64badge/README.txt
