--- boards/hc/mcs92s12ne64/demo9s12ne64/configs/ostest/defconfig
@@ -1,53 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_DEV_CONSOLE is not set
-# CONFIG_STANDARD_SERIAL is not set
-CONFIG_ARCH="hc"
-CONFIG_ARCH_BOARD="demo9s12ne64"
-CONFIG_ARCH_BOARD_DEMOS92S12NEC64=y
-CONFIG_ARCH_CHIP="mcs92s12ne64"
-CONFIG_ARCH_CHIP_MCS92S12NEC64=y
-CONFIG_ARCH_HC=y
-CONFIG_ARCH_STACKDUMP=y
-CONFIG_BOARD_LOOPSPERMSEC=5483
-CONFIG_DISABLE_ENVIRON=y
-CONFIG_DISABLE_MOUNTPOINT=y
-CONFIG_DISABLE_MQUEUE=y
-CONFIG_DISABLE_POSIX_TIMERS=y
-CONFIG_DISABLE_PTHREAD=y
-CONFIG_HCS12_NONBANKED=y
-CONFIG_HCS12_SCI0=y
-CONFIG_HCS12_SERIALMON=y
-CONFIG_IDLETHREAD_STACKSIZE=256
-CONFIG_MAX_TASKS=8
-CONFIG_MAX_WDOGPARMS=2
-CONFIG_MOTOROLA_SREC=y
-CONFIG_NFILE_DESCRIPTORS=3
-CONFIG_NFILE_STREAMS=0
-CONFIG_NUNGET_CHARS=0
-CONFIG_PREALLOC_TIMERS=0
-CONFIG_PREALLOC_WDOGS=4
-CONFIG_PTHREAD_STACK_DEFAULT=256
-CONFIG_RAM_SIZE=65536
-CONFIG_RAM_START=0x20000000
-CONFIG_RR_INTERVAL=200
-CONFIG_SCI0_RXBUFSIZE=32
-CONFIG_SCI0_SERIAL_CONSOLE=y
-CONFIG_SCI0_TXBUFSIZE=32
-CONFIG_SDCLONE_DISABLE=y
-CONFIG_START_DAY=11
-CONFIG_START_MONTH=12
-CONFIG_START_YEAR=2009
-CONFIG_STDIO_BUFFER_SIZE=0
-CONFIG_TASK_NAME_SIZE=0
-CONFIG_TESTING_OSTEST=y
-CONFIG_TESTING_OSTEST_NBARRIER_THREADS=3
-CONFIG_TESTING_OSTEST_STACKSIZE=512
-CONFIG_USERMAIN_STACKSIZE=512
-CONFIG_USER_ENTRYPOINT="ostest_main"
-CONFIG_WDOG_INTRESERVE=0
