--- boards/hc/mcs92s12ne64/ne64badge/src/m9s12_spi.c
@@ -1,104 +1,0 @@
-/****************************************************************************
- * boards/hc/mcs92s12ne64/ne64badge/src/m9s12_spi.c
- *
- *   Copyright (C) 2011 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <stdint.h>
-#include <stdbool.h>
-#include <debug.h>
-
-#include <nuttx/spi/spi.h>
-#include <arch/board/board.h>
-
-#include "ne64badge.h"
-
-#if defined(CONFIG_HCS12_SPI)
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: hcs12_spidev_initialize
- *
- * Description:
- *   Called to configure SPI chip select GPIO pins for the NE64 Badge board.
- *
- ****************************************************************************/
-
-void weak_function hcs12_spidev_initialize(void)
-{
-}
-
-/****************************************************************************
- * Name:  hcs12_spiselect and hcs12_spistatus
- *
- * Description:
- *   The external functions, hcs12_spiselect and hcs12_spistatus must be
- *   provided by board-specific logic.  They are implementations of the select
- *   and status methods of the SPI interface defined by struct spi_ops_s (see
- *   include/nuttx/spi/spi.h).
- *   All other methods (including hcs12_spibus_initialize())
- *   are provided by common HCS12 logic.  To use this common SPI logic on your
- *   board:
- *
- *   1. Provide logic in hcs12_boardinitialize() to configure SPI chip select
- *      pins.
- *   2. Provide hcs12_spiselect() and hcs12_spistatus() functions in your
- *      board-specific logic.  These functions will perform chip selection and
- *      status operations using GPIOs in the way your board is configured.
- *   3. Add a calls to hcs12_spibus_initialize() in your low level application
- *      initialization logic
- *   4. The handle returned by hcs12_spibus_initialize() may then be used to
- *      bind the SPI driver to higher level logic (e.g., calling
- *      mmcsd_spislotinitialize(), for example, will bind the SPI driver to
- *      the SPI MMC/SD driver).
- *
- ****************************************************************************/
-
-void hcs12_spiselect(FAR struct spi_dev_s *dev, uint32_t devid, bool selected)
-{
-}
-
-uint8_t hcs12_spistatus(FAR struct spi_dev_s *dev, uint32_t devid)
-{
-  return SPI_STATUS_PRESENT;
-}
-
-#endif /* CONFIG_HCS12_SPI */
